requirements.txt
quart
tortoise-orm[asyncpg]
tortoise-orm[accel]
uvloop
hypercorn[uvloop]
email_validator
rsa
pyotp
qrcode
loguru
json2html
humanize
passlib
quart-auth
python-dotenv